py-basis/pyproject.toml
[project]
name = "py-basis"
version = "0.1.0"
description = "Binance Basis Monitor - Python Implementation"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "numpy>=1.24.0",
    "websockets>=12.0",
    "orjson>=3.9.10",
    "polars>=0.20.0",
    "rich>=13.7.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.6.0",
]
